Created the ConversionFragment.kt class
