--- app/src/main/java/com/example/android/cryptoapp/ui/ConversionFragment.kt
@@ -1,31 +1,0 @@
-package com.example.android.cryptoapp.ui
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.android.cryptoapp.R
-
-/**
- * A simple [Fragment] subclass.
- * Use the [ConversionFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-class ConversionFragment : Fragment() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-    }
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_conversion, container, false)
-    }
-
-
-
-
-}
