+++ app/src/main/res/layout/fragment_conversion.xml
@@ -14,7 +14,9 @@
         android:layout_height="wrap_content"
         android:orientation="horizontal"
         android:layout_marginBottom="20dp"
-        android:layout_marginTop="20dp">
+        android:layout_marginTop="20dp"
+
+        >
 
 
         <RelativeLayout
